+++ src/PSRule/Parser/RuleLexer.cs
@@ -49,18 +49,15 @@ namespace PSRule.Parser
                         Description(stream, doc) ||
                         Recommendation(stream, doc) ||
                         Notes(stream, doc) ||
-                        RelatedLinks(stream, doc);
+                        Links(stream, doc);
 
                     if (matching)
-                    {
                         continue;
-                    }
                 }
 
                 // Skip the current token
                 stream.Next();
             }
-
             return doc;
         }
 
@@ -70,13 +67,10 @@ namespace PSRule.Parser
         private bool Synopsis(TokenStream stream, RuleDocument doc)
         {
             if (!IsHeading(stream.Current, RULE_ENTRIES_HEADING_LEVEL, DocumentStrings.Synopsis))
-            {
                 return false;
-            }
 
             doc.Synopsis = TextBlock(stream);
             stream.SkipUntilHeader();
-
             return true;
         }
 
@@ -86,13 +80,10 @@ namespace PSRule.Parser
         private bool Description(TokenStream stream, RuleDocument doc)
         {
             if (!IsHeading(stream.Current, RULE_ENTRIES_HEADING_LEVEL, DocumentStrings.Description))
-            {
                 return false;
-            }
 
             doc.Description = TextBlock(stream);
             stream.SkipUntilHeader();
-
             return true;
         }
 
@@ -102,13 +93,10 @@ namespace PSRule.Parser
         private bool Recommendation(TokenStream stream, RuleDocument doc)
         {
             if (!IsHeading(stream.Current, RULE_ENTRIES_HEADING_LEVEL, DocumentStrings.Recommendation))
-            {
                 return false;
-            }
 
             doc.Recommendation = TextBlock(stream);
             stream.SkipUntilHeader();
-
             return true;
         }
 
@@ -118,33 +106,28 @@ namespace PSRule.Parser
         private bool Notes(TokenStream stream, RuleDocument doc)
         {
             if (!IsHeading(stream.Current, RULE_ENTRIES_HEADING_LEVEL, DocumentStrings.Notes))
-            {
                 return false;
-            }
 
             doc.Notes = TextBlock(stream);
             stream.SkipUntilHeader();
-
             return true;
         }
 
-        private bool RelatedLinks(TokenStream stream, RuleDocument doc)
+        /// <summary>
+        /// Read links.
+        /// </summary>
+        private bool Links(TokenStream stream, RuleDocument doc)
         {
             if (!IsHeading(stream.Current, RULE_ENTRIES_HEADING_LEVEL, DocumentStrings.Links))
-            {
                 return false;
-            }
 
-            List<Link> links = new List<Link>();
-
+            var links = new List<Link>();
             stream.Next();
-
-            while (stream.IsTokenType(MarkdownTokenType.Link, MarkdownTokenType.LinkReference, MarkdownTokenType.LineBreak))
+            while (stream.IsTokenType(MarkdownTokenType.Link, MarkdownTokenType.LinkReference, MarkdownTokenType.LineBreak, MarkdownTokenType.Text))
             {
-                if (stream.IsTokenType(MarkdownTokenType.LineBreak))
+                if (stream.IsTokenType(MarkdownTokenType.LineBreak) || stream.IsTokenType(MarkdownTokenType.Text))
                 {
                     stream.Next();
-
                     continue;
                 }
 
@@ -160,37 +143,28 @@ namespace PSRule.Parser
                     var target = stream.ResolveLinkTarget(link.Uri);
                     link.Uri = target.Text;
                 }
-
                 links.Add(link);
-
                 stream.Next();
             }
-
             stream.SkipUntilHeader();
-
             doc.Links = links.ToArray();
-
             return true;
         }
 
         private TextBlock TextBlock(TokenStream stream)
         {
             var useBreak = stream.Current.IsDoubleLineEnding();
-
             stream.Next();
-
             var text = ReadText(stream);
-
             return new TextBlock(text: text, formatOption: useBreak ? FormatOptions.LineBreak : FormatOptions.None);
         }
 
         /// <summary>
         /// Read tokens from the stream as text.
         /// </summary>
-        private string ReadText(TokenStream stream, bool includeNonYamlFencedBlocks = false)
+        private static string ReadText(TokenStream stream, bool includeNonYamlFencedBlocks = false)
         {
             var sb = new StringBuilder();
-
             while (stream.IsTokenType(MarkdownTokenType.Text, MarkdownTokenType.Link, MarkdownTokenType.FencedBlock, MarkdownTokenType.LineBreak))
             {
                 if (stream.IsTokenType(MarkdownTokenType.Text))
@@ -202,16 +176,12 @@ namespace PSRule.Parser
                 {
                     AppendEnding(sb, stream.Peak(-1));
                     sb.Append(stream.Current.Meta);
-
                     if (!string.IsNullOrEmpty(stream.Current.Text))
-                    {
                         sb.AppendFormat(" ({0})", stream.Current.Text);
-                    }
                 }
                 else if (stream.IsTokenType(MarkdownTokenType.LinkReference))
                 {
                     AppendEnding(sb, stream.Peak(-1));
-
                     sb.Append(stream.Current.Meta);
                 }
                 else if (stream.IsTokenType(MarkdownTokenType.FencedBlock))
@@ -220,52 +190,37 @@ namespace PSRule.Parser
                     if (!includeNonYamlFencedBlocks || string.Equals(stream.Current.Meta, "yaml", StringComparison.OrdinalIgnoreCase))
                     {
                         if (stream.PeakTokenType(-1) == MarkdownTokenType.LineBreak)
-                        {
                             AppendEnding(sb, stream.Peak(-1));
-                        }
 
                         break;
                     }
-
                     AppendEnding(sb, stream.Peak(-1), preserveEnding: true);
                     sb.Append(stream.Current.Text);
                 }
                 else if (stream.IsTokenType(MarkdownTokenType.LineBreak))
-                {
                     AppendEnding(sb, stream.Peak(-1));
-                }
 
                 stream.Next();
             }
 
             if (stream.EOF && stream.Peak(-1).Flag.HasFlag(MarkdownTokens.Preserve) && stream.Peak(-1).Flag.HasFlag(MarkdownTokens.LineEnding))
-            {
                 AppendEnding(sb, stream.Peak(-1));
-            }
 
             return sb.ToString();
         }
 
-        private void AppendEnding(StringBuilder stringBuilder, MarkdownToken token, bool preserveEnding = false)
+        private static void AppendEnding(StringBuilder stringBuilder, MarkdownToken token, bool preserveEnding = false)
         {
             if (token == null || stringBuilder.Length == 0 || !token.Flag.IsEnding())
-            {
                 return;
-            }
 
             if (!preserveEnding && token.Flag.ShouldPreserve())
-            {
                 preserveEnding = true;
-            }
 
             if (token.IsDoubleLineEnding())
-            {
                 stringBuilder.Append(preserveEnding ? string.Concat(Environment.NewLine, Environment.NewLine) : Environment.NewLine);
-            }
             else if (token.IsSingleLineEnding())
-            {
                 stringBuilder.Append(preserveEnding ? Environment.NewLine : Space);
-            }
         }
     }
 }