+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Fixed JSON de-serialization fails with single object. [#379](https://github.com/Microsoft/PSRule/issues/379)
 - Fixed stack overflow when parsing malformed JSON. [#380](https://github.com/Microsoft/PSRule/issues/380)
+- Added rule documentation links and notes to help. [#382](https://github.com/Microsoft/PSRule/issues/382)
+  - Added `-Full` switch to `Get-PSRuleHelp` to display links and notes sections.
 
 ## v0.13.0-B1912043 (pre-release)